Split settings navigation metadata
