+++ src/renderer/src/hooks/settings-navigation-build-options.ts
@@ -1,0 +1,14 @@
+import type { Repo } from '../../../shared/repo-types'
+
+export type SettingsNavigationBuildOptions = {
+  isMac: boolean
+  isWindows: boolean
+  isLocalWindowsHost: boolean
+  isWindowsTerminalHost: boolean
+  isWebClient: boolean
+  managedBrowserCreationEnabled: boolean
+  mobileEmulatorCreationEnabled: boolean
+  isDev: boolean
+  isLinearConnected: boolean
+  repos: readonly Repo[]
+}
